requirements.txt
fastapi
redis
uvicorn
python-dotenv
httpx
openai
stripe
motor